pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text-extractor"
version = "2.0.0"
description = "Extract text from PDF, DOCX, ODT, and TXT files in a directory tree and save to JSON"
readme = "README.md"
license = {text = "MIT License"}
authors = [
    {name = "dpmattos", email = "dpmattos@uwaterloo.ca"}
]
maintainers = [
    {name = "dpmattos", email = "dpmattos@uwaterloo.ca"}
]
keywords = ["pdf", "docx", "odt", "txt", "extraction", "text", "json", "docopt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]
requires-python = ">=3.7"
dependencies = [
    "docopt-ng>=0.8.0",
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "odfpy>=1.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
speed = [
    "pypdf>=3.0.0",
]
ocr = [
    "pdf2image>=1.16.0",
    "pytesseract>=0.3.10",
]
all = [
    "pypdf>=3.0.0",
    "pdf2image>=1.16.0",
    "pytesseract>=0.3.10",
]

[project.urls]
Homepage = "https://github.com/dpmattos/text-extractor"
Documentation = "https://github.com/dpmattos/text-extractor#readme"
Repository = "https://github.com/dpmattos/text-extractor.git"
Issues = "https://github.com/dpmattos/text-extractor/issues"

[project.scripts]
text-extract = "text_extractor:main"

[tool.setuptools]
py-modules = ["text_extractor"]

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", "*.egg-info"]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false
strict_optional = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=text_extractor --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["text_extractor.py"]
omit = ["tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
]